requirements.txt
bokeh==3.1.1
colorama==0.4.6
contourpy==1.0.5
cycler==0.11.0
filelock==3.13.1
fonttools==4.51.0
fsspec==2024.2.0
fuzzycmeans==1.0.4
importlib-resources==6.1.1
intel-openmp==2021.4.0
Jinja2==3.1.3
joblib==1.4.2
kiwisolver==1.4.4
MarkupSafe==2.1.5
matplotlib==3.7.5
mpmath==1.3.0
networkx==3.0
numpy==1.24.1
packaging==23.2
pandas==2.0.3
pillow==10.2.0
pip==23.3.1
ply==3.11
pyparsing==3.0.9
PyQt5==5.15.10
PyQt5-sip==12.13.0
python-dateutil==2.8.2
pytz==2024.1
PyYAML==6.0.1
scikit-fuzzy==0.4.2
scikit-learn==1.3.2
scipy==1.10.1
seaborn==0.13.2
setuptools==69.5.1
sip==6.7.12
six==1.16.0
sympy==1.12
tbb==2021.11.0
threadpoolctl==3.5.0
tomli==2.0.1
tornado==6.3.3
tqdm==4.66.2
typing_extensions==4.9.0
tzdata==2024.1
unicodedata2==15.1.0
wheel==0.43.0
xyzservices==2024.6.0
zipp==3.17.0
